requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.11.7
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
bandit==1.7.4
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
commonmark==0.9.1
cryptography==37.0.4
cycler==0.11.0
dateparser==1.1.1
debugpy==1.6.2
decorator==5.1.1
deep-translator==1.8.3
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.5
docutils==0.19
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.1
filelock==3.7.1
flake8==4.0.1
fonttools==4.34.4
frozenlist==1.3.0
gensim==4.2.0
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
isort==5.10.1
jedi==0.18.1
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
keyring==23.7.0
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
multidict==6.0.2
mypy==0.961
mypy-extensions==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.12
numpy==1.23.1
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pbr==5.9.0
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipenv==2022.7.4
pkginfo==1.8.3
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.12.0
pylama==8.3.8
pylint==2.14.5
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-binance==1.0.16
python-dateutil==2.8.2
pytoolconfig==1.2.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==23.2.0
readme-renderer==35.0
regex==2022.3.2
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.5.1
rope==1.2.0
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
SecretStorage==3.3.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smart-open==6.0.0
smmap==5.0.0
snowballstemmer==2.2.0
snscrape==0.4.3.20220106
soupsieve==2.3.2.post1
stack-data==0.3.0
stevedore==4.0.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
twine==4.0.1
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
ujson==5.4.0
urllib3==1.26.10
virtualenv==20.15.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.3
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.1
